requirements.txt
aiokafka==0.7.0
alembic==1.5.4
arrow==0.17.0
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.4
cryptography==3.1.1
Faker==6.1.1
fastapi==0.63.0
flake8==3.8.4
h11==0.12.0
humps==0.2.2
idna==2.10
iniconfig==1.1.1
isort==5.7.0
kafka-python==2.0.2
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
phonenumbers==8.12.18
pluggy==0.13.1
protobuf==3.14.0
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-cov==2.11.1
pytest-flake8==1.0.7
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
uvicorn==0.13.3
wrapt==1.12.1
